Add PortalProductMapper class with mapping functionality
The new PortalProductMapper class allows for the mapping of products by name. It includes methods that prepare product name and product-by-name maps, as well as a method to map products provided specific product attributes. An additional MappingResult inner class is included to handle the return of product mappings and potential errors.