Add LogEntryEventStreamController class and corresponding tests
This commit introduces the LogEntryEventStreamController class, used to stream Log Entries in console and Tabular view. The class includes methods to get and enable streaming log entries. In addition, it integrates with LoggerParameter to enable log stream and fetch the display fields. Associated unit tests have been developed to validate the functionality. Modifications to log entries include adding and configuring the logger scenarios, upserting the logs, and appending rule-based tags.